-data, -index

+++ xml/xml.cpp
@@ -273,6 +273,8 @@ void Schema::merge (Schema& other)
 
 
 void Schema::printTableDdl (ostream &os,
+                            bool dataP,
+                            bool indexP,
                             const Schema* curTable) const
 { 
   IMPLY (curTable, curTable->multiple);
@@ -280,30 +282,36 @@ void Schema::printTableDdl (ostream &os,
   if (multiple)
   {
     const string table (getColumnName (nullptr));
-    // "xml_num_", "id_" ??
-    os << "drop   table " + table << ";" << endl 
-       << "create table " + table << endl << '(' << endl
-       << "  xml_num_ int  not null" << endl
-       << ", id_ bigint  not null" << endl;  
     const string refTable (curTable ? curTable->getColumnName (nullptr) : "");
-    if (! refTable. empty ())
-      os << ", " << refTable << "_id_ bigint  not null" << endl;  
     curTable = this;
-    printColumnDdl (os, curTable);  
-    os << ");" << endl;
-    os << "alter table " << table << " add constraint " << table << "_pk primary key (xml_num_, id_);" << endl;
-    os << "grant select on " << table << " to public;" << endl;
-    if (! refTable. empty ())
+    if (dataP)
     {
-      os << "create index " << table << "_idx on " << table << "(xml_num_," << refTable << "_id_);" << endl;
-      os << "alter table " << table << " add constraint " << table << "_fk foreign key (xml_num_, " << refTable << "_id_) references " << refTable << "(xml_num_,id_) on delete cascade;" << endl;
-      os << "-- delete /*top (1000000)*/ from " << table << " with (tablock) where not exists (select null from " << refTable << " (nolock) where " << refTable << ".xml_num_ = " << table << ".xml_num_ and " << refTable << ".id_ = " << table << "." << refTable << "_id_);" << endl;
+      // "xml_num_", "id_" ??
+      os << "drop   table " + table << ";" << endl 
+         << "create table " + table << endl << '(' << endl
+         << "  xml_num_ int  not null" << endl
+         << ", id_ bigint  not null" << endl;  
+      if (! refTable. empty ())
+        os << ", " << refTable << "_id_ bigint  not null" << endl;  
+      printColumnDdl (os, curTable);  
+      os << ");" << endl;
+      os << "grant select on " << table << " to public;" << endl;
+    }
+    if (indexP)
+    {
+      os << "alter table " << table << " add constraint " << table << "_pk primary key (xml_num_, id_);" << endl;
+      if (! refTable. empty ())
+      {
+        os << "create index " << table << "_idx on " << table << "(xml_num_," << refTable << "_id_);" << endl;
+        os << "alter table " << table << " add constraint " << table << "_fk foreign key (xml_num_, " << refTable << "_id_) references " << refTable << "(xml_num_,id_) on delete cascade;" << endl;
+        os << "-- delete /*top (1000000)*/ from " << table << " with (tablock) where not exists (select null from " << refTable << " (nolock) where " << refTable << ".xml_num_ = " << table << ".xml_num_ and " << refTable << ".id_ = " << table << "." << refTable << "_id_);" << endl;
+      }
     }
     os << endl << endl;
   }
 
   for (const auto& it : name2schema)
-    it. second->printTableDdl (os, curTable);
+    it. second->printTableDdl (os, dataP, indexP, curTable);
 }
